da big bang

+++ Age of war/src/Draws/BasicProjectileDraw.java
@@ -10,7 +10,7 @@ public class BasicProjectileDraw extends Drawable{
 	private int screenHeight;
 	private int screenWidth;
 	
-	public BasicProjectileDraw(double xPos, double yPos, int screenHeight, int screenWidth) {
+	public BasicProjectileDraw(float xPos, float yPos, int screenHeight, int screenWidth) {
 		super(xPos, yPos, screenHeight, screenWidth);
 		//this.screenHeight = leftTexture.getHeight() * (Game.HEIGHT * Game.SCALE)/Gdx.graphics.getHeight()/scale;
 		//this.screenWidth = leftTexture.getWidth()* (Game.WIDTH * Game.SCALE)/Gdx.graphics.getWidth()/scale;
