+++ Age of war/src/Draws/Drawable.java
@@ -18,7 +18,7 @@ public abstract class Drawable {
 	protected int screenHeight;
 	protected int currentSprite;
 	
-	public Drawable(double xPos2, double yPos2, int screenHeight, int screenWidth){
+	public Drawable(float xPos2, float yPos2, int screenHeight, int screenWidth){
 		
 		this.xPos = xPos2;
 		this.yPos = yPos2;
@@ -26,7 +26,7 @@ public abstract class Drawable {
 		this.screenWidth = screenWidth;
 	}
 	
-	public Drawable(double xPos2, double yPos2){
+	public Drawable(float xPos2, float yPos2){
 		
 		this.xPos = xPos2;
 		this.yPos = yPos2;
