+++ Age of war/src/Draws/GroundDraw.java
@@ -13,7 +13,7 @@ public class GroundDraw extends Drawable{
 	private int screenHeight;
 	private int screenWidth;
 	
-	public GroundDraw(double xPos, double yPos) {
+	public GroundDraw(float xPos, float yPos) {
 		super(xPos, yPos);
 		this.screenHeight = Game.GROUND_HEIGHT * Gdx.graphics.getHeight() / (Game.HEIGHT * Game.SCALE);
 		this.screenWidth = Gdx.graphics.getWidth();
